% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binarize}
\alias{binarize}
\title{binarization function
1 if larger than threshold, otherwise 0}
\usage{
binarize(x, th = 0)
}
\arguments{
\item{x}{numeric object}

\item{th}{threshold, a numeric value}
}
\value{
binarized object
}
\description{
binarization function
1 if larger than threshold, otherwise 0
}
\examples{

}
