% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bitr_id2symbol}
\alias{bitr_id2symbol}
\title{Convert ID to symbol use clusterProfiler::bitr}
\usage{
bitr_id2symbol(geneID, OrgDb = "org.Hs.eg.db", drop = T)
}
\arguments{
\item{geneID}{input gene id}

\item{OrgDb}{annotation db}

\item{drop}{drop NA or not}
}
\value{
data.frame of id and symbol map
}
\description{
Convert ID to symbol use clusterProfiler::bitr
}
\examples{

}
