% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcDetectionRate}
\alias{calcDetectionRate}
\title{calc mean and rate of gene expression}
\usage{
calcDetectionRate(data, group.by)
}
\arguments{
\item{data}{gene by cell matrix}

\item{group.by}{vector of cell attribute}
}
\value{
data.frame of two columns
}
\description{
calc mean and rate of gene expression
}
\examples{

}
