% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcFoldChange}
\alias{calcFoldChange}
\title{calculate Fold Change of gene expression values}
\usage{
calcFoldChange(
  x,
  y,
  log.base = exp(1),
  exact = T,
  exact.plus = 1,
  return.log = F
)
}
\arguments{
\item{x, y}{gene expression values in log scale}

\item{log.base}{base of log scale}

\item{exact}{usually gene expression values were scaled and plus a pseudo count (eg. 1) and then log transformed. represented as log(UMI/10+1). Fold change would be not exact if do not minus the pseudo count, although pseudo count is usefull when denominator is zero.}

\item{exact.plus}{pseudo count, usualy be 1.}

\item{return.log}{fold change value returned in log scale or not}
}
\value{
fold change value of x compared to y
}
\description{
calculate Fold Change of gene expression values
}
\examples{

}
