% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcLogMean}
\alias{calcLogMean}
\title{calculate mean of gene expression values}
\usage{
calcLogMean(x, log.base = exp(1), return.log = T)
}
\arguments{
\item{x}{gene expression values in log scale}

\item{log.base}{base of log scale}

\item{return.log}{mean returned in log scale or not}
}
\value{
mean value
}
\description{
calculate mean of gene expression values
}
\examples{

}
