% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckBrewerPal}
\alias{CheckBrewerPal}
\title{Check pal in brewer}
\usage{
CheckBrewerPal(pal = "YlGn", n = 5)
}
\arguments{
\item{pal}{a string of names of brewer.pal, or a vector of colors to be used to produce n gradient colors.}

\item{n}{length of returned color vector}
}
\value{
a vector of colors
}
\description{
Check pal in brewer
}
