% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckDuplicate}
\alias{CheckDuplicate}
\title{Check Function}
\usage{
CheckDuplicate(x, strict = T, rm_dup = T)
}
\arguments{
\item{x}{a vector}

\item{strict}{if strict, duplicated x will result in stop.}
}
\value{
deduplicated x
}
\description{
Check Function
}
\examples{

}
