% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckFeatures}
\alias{CheckFeatures}
\title{Check Feature Duplicate and Not In Data}
\usage{
CheckFeatures(object, features, rm_dup = T, rm_notin = T)
}
\arguments{
\item{object}{data or seurat3 object}

\item{features}{features}

\item{rm_dup}{remove duplicated elements}

\item{rm_notin}{remove features not in data}
}
\value{
updated features
}
\description{
Check Feature Duplicate and Not In Data
}
\examples{

}
