% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chiseq_roe}
\alias{chiseq_roe}
\title{Ro/e score analysis}
\usage{
chiseq_roe(tab, ...)
}
\arguments{
\item{tab}{table or matrix}

\item{...}{other params passed to chisq.test}
}
\value{
Ro/e
}
\description{
Ro/e score analysis
}
\examples{

}
