% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color2RGB}
\alias{color2RGB}
\title{convert a color name to RGB character}
\usage{
color2RGB(col, alpha = 255)
}
\arguments{
\item{col}{color name}

\item{alpha}{transparent, 0 to 255}
}
\value{
a character
}
\description{
convert a color name to RGB character
}
\examples{

}
