% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CSI}
\alias{CSI}
\title{Connection Specificity Index (CSI)
Connection Specificity Index (CSI) was defined as the fraction of TFs connected to a and b that have a PCC lower than the PCC of a and b;}
\usage{
CSI(pccMat, shrink = 0.05)
}
\arguments{
\item{pccMat}{Correaltion Coefficient Matrix, a full rank square matrix.}

\item{shrink}{shink the PCC threshold for each element}
}
\value{
CSI Matrix
}
\description{
Connection Specificity Index (CSI)
Connection Specificity Index (CSI) was defined as the fraction of TFs connected to a and b that have a PCC lower than the PCC of a and b;
}
\details{
see http://csbio.cs.umn.edu/similarity_index/guide.php
https://www.nature.com/articles/nmeth.2728
https://doi.org/10.1016/j.celrep.2018.10.045
}
\examples{

}
