% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df2list}
\alias{df2list}
\title{to list}
\usage{
df2list(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{vector or data frame containing values to be divided into groups.}

\item{f}{a \sQuote{factor} in the sense that \code{\link[base]{as.factor}(f)}
    defines the grouping, or a list of such factors in which case their
    interaction is used for the grouping.}

\item{drop}{logical indicating if levels that do not occur should be dropped
    (if \code{f} is a \code{factor} or a list).}

\item{...}{further potential arguments passed to methods.}
}
\value{

}
\description{
A alias function of base::split
}
\examples{

}
