% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{displayBrewers}
\alias{displayBrewers}
\title{Show colours}
\usage{
displayBrewers(
  n = NULL,
  type = "all",
  select = NULL,
  exact.n = TRUE,
  colorblindFriendly = FALSE
)
}
\arguments{
\item{n}{Number of different colors in the palette, minimum 3, 
maximum depending on palette}

\item{type}{One of the string "div", "qual", "seq", or "all" }

\item{select}{A list of names of existing palettes}

\item{exact.n}{If TRUE, only display palettes with a color number given by n}

\item{colorblindFriendly}{if TRUE, display only colorblind friendly palettes}
}
\description{
short for RColorBrewer::display.brewer.all
}
