% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropColors}
\alias{dropColors}
\title{drop unused colors}
\usage{
dropColors(colors, df, drop_col = NULL)
}
\arguments{
\item{colors}{a named list of color vectors}

\item{df}{data.frame to check}

\item{drop_col}{drop which color. NULL is all.}
}
\value{
a list of updated colors
}
\description{
This may be usefull for annotation_colors in pheatmap or pl_heatmap
}
\examples{

}
