% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportUMIByBatch}
\alias{exportUMIByBatch}
\title{export UMI to each file by each Batch}
\usage{
exportUMIByBatch(umi, batch)
}
\arguments{
\item{umi}{umi matrix}

\item{batch}{a vector for cell batch information}
}
\value{
csv files
}
\description{
this may be used for preparation of files to GEO
}
\examples{

}
