% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMGIHomoGeneTable}
\alias{getMGIHomoGeneTable}
\title{Generate Homologous Gene ID Table from MGI's HOM_AllOrganism.rpt file}
\usage{
getMGIHomoGeneTable(
  MGI_file,
  colnames_keys = c("HomoloGene ID", "Common Organism Name", "Symbol"),
  multiple_orgs = c("human", "mouse, laboratory", "macaque, rhesus"),
  multiple_orgs_short = c("Hs", "Mm", "Mc"),
  centric_org = "human"
)
}
\arguments{
\item{MGI_file}{HOM_AllOrganism.rpt downloaded from http://www.informatics.jax.org/downloads/reports/HOM_AllOrganism.rpt}

\item{colnames_keys}{should not be changed!}

\item{multiple_orgs}{organisms, in \code{Common Organism Name} column of \code{MGI_file}, to be collapsed or compared,}

\item{multiple_orgs_short}{short prefix for the organisms in \code{mulitple_orgs}}

\item{centric_org}{which organism's Symbol should be used as HomoloSymbol}
}
\value{
a data.frame table
}
\description{
Generate Homologous Gene ID Table from MGI's HOM_AllOrganism.rpt file
}
\examples{

}
