% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gseKEGG_symbol}
\alias{gseKEGG_symbol}
\title{Gene Set Enrichment Analysis of KEGG}
\usage{
gseKEGG_symbol(
  geneList,
  organism = "hsa",
  keyType = "symbol",
  exponent = 1,
  nPerm = 1000,
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  verbose = TRUE,
  use_internal_data = FALSE,
  seed = FALSE,
  by = "fgsea"
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'. For hsa and mmu, "symbol" is also supported.}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}

\item{seed}{logical}

\item{by}{one of 'fgsea' or 'DOSE'}
}
\value{

}
\description{
Gene Set Enrichment Analysis of KEGG
}
\examples{

}
