% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{H}
\alias{H}
\title{Shannon entropy for distribution P}
\usage{
H(p, unit = log2)
}
\arguments{
\item{p}{a discrete distribution, whose elements should sum up to 1.
Elements less than 0 will be ignored, and all elements will be forced to be normalized to sum 1 (p/sum(p)).}

\item{unit}{log, log2, log10 used in calculation of Shannon entropy}
}
\value{
Shannon entropy value: sum(-p*log(p))
}
\description{
Shannon entropy for distribution P
}
\examples{

}
