% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h5sToSeurat}
\alias{h5sToSeurat}
\title{read multiple h5 files into one seurat object}
\usage{
h5sToSeurat(filenames, filemeta, use.names = TRUE, unique.features = TRUE)
}
\arguments{
\item{filenames}{files with full path}

\item{filemeta}{a named list of several vectors to be used as meta.data. list name as colnames, length of each vector should be equal to length of `filenames`}
}
\value{
seurat object
}
\description{
read multiple h5 files into one seurat object
}
\examples{

}
