% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HeaderMesseage}
\alias{HeaderMesseage}
\title{Messages with header}
\usage{
HeaderMesseage(header, messages, ...)
}
\arguments{
\item{header}{header string}

\item{messages}{message vectors}

\item{...}{other params passed to str_c(messages, ...)}
}
\value{
string
}
\description{
Messages with header
}
\examples{

}
