% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{homoMSigDB}
\alias{homoMSigDB}
\title{Conver MSigDB Human genes to Homologous Organism genes}
\usage{
homoMSigDB(
  files = NULL,
  organism = "mouse",
  use.entrez = F,
  use.biomaRt = T,
  hcop.file = NULL,
  try.most = T
)
}
\arguments{
\item{files}{MSigDB files in gmt format}

\item{organism}{conversion organism}

\item{use.entrez}{set TRUE if gmt files use entrez gene id but not symbol}

\item{use.biomaRt}{use BioMart package to do conversion}

\item{hcop.file}{Not Supportted for now! use HGNC Comparison of Orthology Predictions (HCOP) to do conversion. http://www.genenames.org/cgi-bin/hcop}

\item{try.most}{Not used}
}
\value{
files
}
\description{
Conver MSigDB Human genes to Homologous Organism genes
}
\examples{

}
