% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inferNormalMixTh}
\alias{inferNormalMixTh}
\title{EM Algorithm for Mixtures of Univariate Normals}
\usage{
inferNormalMixTh(x, posterior = 0.5, ...)
}
\arguments{
\item{x}{A vector of length n consisting of the data.}

\item{posterior}{threshold of posterior probability. Default: 0.5}

\item{...}{othter parameters passed to mixtools::normalmixEM}
}
\value{
a list of class mixEM defined in mixtools::normalmixEM and items:
predict: a vector of predicted class
threshold: a value
}
\description{
A wrapper of mixtools::normalmixEM with prediction by posterior probability
}
\examples{

}
