% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{InferScaleFactor}
\alias{InferScaleFactor}
\title{Infer the scale factor used for normalization}
\usage{
InferScaleFactor(x, log_base = base::exp(1), pseudo_count = 1)
}
\arguments{
\item{x}{matrix or data.frame}

\item{log_base}{normalization log scaled base}

\item{pseudo_count}{usually 1.}
}
\value{

}
\description{
Infer the scale factor used for normalization
}
