% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inferUMI}
\alias{inferUMI}
\title{Infer UMI from scale-factor-normalized data}
\usage{
inferUMI(data, base = exp(1))
}
\arguments{
\item{data}{scale-factor-normalized data (normalized by column)}
}
\value{
a matrix of inferred UMI data
}
\description{
Pseudo: the smallest non-zero value (normalized) is corresponding to 1 of UMI.
}
\examples{

}
