% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{JSD2}
\alias{JSD2}
\title{JSD for two vector}
\usage{
JSD2(p, q, pseudocount = 1e-16, unit = log2)
}
\arguments{
\item{p}{a discrete distribution}

\item{q}{a discrete distribution}

\item{pseudocount}{in case of elements of zero}

\item{unit}{log, log2, log10 used in calculation of Shannon entropy}
}
\value{
Jensen Shanonn Divergence
}
\description{
JSD for two vector
}
\examples{

}
