% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{JSD_group}
\alias{JSD_group}
\title{Group specific distance}
\usage{
JSD_group(data, group, unit = log2)
}
\arguments{
\item{data}{a matrix, for example, gene expression, regulon scores ...}

\item{group}{a category vector}

\item{unit}{log, log2, log10 used in calculation of Shannon entropy}
}
\value{
Similarity data for each feature (row in data) and each category (group);
divergence: Jensen Shanonn Divergence;
distance: sqrt(divergence);
specificity: specificity score;
group: group with max specificity score;
}
\description{
Group specific distance
}
\examples{
# for cell-type specific score
data <- matrix(1:100, 10)
group = sample(1:3, 10, replace = T)
JSD_group(data, group)

}
