% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2df}
\alias{list2df}
\title{convert a list of vectors with different lenght to data.frame}
\usage{
list2df(x)
}
\arguments{
\item{x}{a list of vectors}
}
\value{
a data.frame
data.frame is easier to view/export to excel
}
\description{
convert a list of vectors with different lenght to data.frame
}
\examples{

}
