% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeSeuratList}
\alias{mergeSeuratList}
\title{merge seurat list}
\usage{
mergeSeuratList(object_list, genes_used = NULL)
}
\arguments{
\item{object_list}{a list of seurat objects}

\item{genes_used}{if NULL use their shared genes}
}
\value{
one seurat object
}
\description{
merge seurat list
}
\examples{

}
