% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Message}
\alias{Message}
\title{format a text message}
\usage{
Message(head = NULL, items = NULL, item_sep = ", ", tail = "\\n")
}
\arguments{
\item{head}{head of message}

\item{items}{join multiple `items` into a single string with `item_sep` as seperator.}

\item{item_sep}{item seperator}

\item{tail}{tail of message. `\n`` should be used at the end always.}
}
\value{
a string
}
\description{
format a text message
}
