% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_averageHeatmap}
\alias{pl_averageHeatmap}
\title{Heatmap of average expression of select features}
\usage{
pl_averageHeatmap(
  data,
  features,
  group_by,
  groups = NULL,
  method = base::mean,
  by_log = F,
  log_base = base::exp(1),
  exclude_zero = F,
  outlier_cutoff = 1,
  heat_color = c("darkblue", "royalblue", "grey100", "tomato", "brown"),
  scale_feature = T,
  cap_max = 2,
  cap_min = -2,
  cluster_rows = F,
  cluster_cols = F,
  show_rownames = F,
  ...
)
}
\arguments{
\item{data}{expression matrix of feature by cell, ususally normalized (in log scale) but not feature scaled.}

\item{features}{features to be visualized}

\item{group_by}{a vector including cell annotations}

\item{groups}{only these selected groups to be visualized}

\item{method}{using `base::mean` or `stats::median` or your custom function for each feature vector;}

\item{by_log}{calculate by log scale ; If `TRUE` log_base will be ignored.}

\item{log_base}{`base::exp(1)` or `2` or other else, depends on your normalization method.}

\item{exclude_zero}{exclude zeros when calculate average/median feature value. Note exclude_zero first, outlier_cutoff second.}

\item{outlier_cutoff}{sometimes outliers (several extremely high cells) should be excluded when do summarise. Set `0.99` to exclude top 1% cells. (default: 1)}

\item{heat_color}{heatmap colors}

\item{scale_feature}{whether do scale by feature}

\item{cap_max}{if scale_feature, set cap for max}

\item{cap_min}{if scale_feature, set cap for min}

\item{cluster_rows}{row clustering}

\item{cluster_cols}{column clustering}

\item{show_rownames}{show rownames}

\item{...}{}
}
\value{
a pheatmap object
}
\description{
Heatmap of average expression of select features
}
