% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_dotplot_deg}
\alias{pl_dotplot_deg}
\title{DotPlot of top n degs}
\usage{
pl_dotplot_deg(object, degs, topn = 5, group_by = NULL, ...)
}
\arguments{
\item{object}{seurat object}

\item{degs}{degs data.farme derived from `FindAllMarkers`}

\item{topn}{integer,}

\item{group_by}{vector}

\item{...}{other params passed to `pl_dotplot`}
}
\value{
a plot by pl_dotplot
}
\description{
DotPlot of top n degs
}
\examples{

}
