% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_genePatternDiagram}
\alias{pl_genePatternDiagram}
\title{Visualize gene expression pattern using boxVlnPlot and smooth line}
\usage{
pl_genePatternDiagram(mat, cell_group, gene_group, ncol = 1, return_data = F)
}
\arguments{
\item{mat}{gene x cell matrix}

\item{cell_group}{a vector of cells' group corresponding to colnames(mat)}

\item{gene_group}{a vector of genes' group corresponding to rownames(mat)}

\item{return_data}{if TRUE return summarised data, otherwise return ggplot}
}
\value{
ggplot or data.frame
}
\description{
Visualize gene expression pattern using boxVlnPlot and smooth line
}
\examples{

}
