% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_genePatternHeatmap}
\alias{pl_genePatternHeatmap}
\title{visualize genes patterns heatmap using kmeans clustering}
\usage{
pl_genePatternHeatmap(
  mat,
  genes_select,
  cellAnnot,
  geneAnnot,
  cluster_colname = "identity",
  clusters_display,
  clusters_stat = NULL,
  filter_gene_by_stat = F,
  stat_pvalue = 0.05,
  stat_padj = 1,
  do_gene_kmeans = T,
  kmeans_k = 10,
  kmeans_seed = 1,
  gene_scale = T,
  min = -2,
  max = 2,
  display_gaps = c("row", "col"),
  return_data = F,
  ...
)
}
\arguments{
\item{mat}{gene x cell expression matrix. cell order is kept for heatmap.}

\item{genes_select}{genes to be display}

\item{cellAnnot}{cell annotation data frame}

\item{geneAnnot}{gene annotation, used for annotation bar in heatmap.}

\item{cluster_colname}{which column to be used for group}

\item{clusters_display}{subset clusters to be displayed}

\item{clusters_stat}{subset clusters to be used for filtering genes by statistical test}

\item{filter_gene_by_stat}{whether to do filter gene}

\item{stat_pvalue}{threshold for filtering genes}

\item{stat_padj}{threshold for filtering genes. both pvalue and padjust will be fullfilment.}

\item{do_gene_kmeans}{use kmeans to determine gene order}

\item{kmeans_k}{k parameter in kmeans}

\item{kmeans_seed}{random seeds for kmeans}

\item{gene_scale}{do gene-wise scale for display}

\item{min}{cap the scaled value}

\item{max}{cap the scaled value}

\item{display_gaps}{do \code{gaps_row} and \code{gaps_col} of pheatmap}

\item{return_data}{return mat and kmeans results}

\item{...}{parameters used in pheatmap}
}
\value{
pheatmap plot
}
\description{
visualize genes patterns heatmap using kmeans clustering
}
\examples{

}
