% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_heatmap}
\alias{pl_heatmap}
\title{plot heatmap using pheatmap}
\usage{
pl_heatmap(
  data,
  features = NULL,
  color = gradient_colors$f_BuWtRd(99, interpolate = "linear"),
  meta_data = NA,
  annot_col = NA,
  permutation_by = NA,
  seed = 666,
  feature_data = NA,
  annot_row = NA,
  smooth_n = NA,
  do_scale = F,
  smooth_lowess = F,
  lowess_span = 2/3,
  cap_max = NA,
  cap_min = NA,
  annot_colors = NA,
  rownames_to_show = NA,
  ...
)
}
\arguments{
\item{data}{feature by cell data matrix}

\item{features}{features to display}

\item{color}{heatmap colors}

\item{meta_data}{cell annotation data frame}

\item{annot_col}{select colnames of meta_data}

\item{permutation_by}{permutate and order the cells}

\item{seed}{permutation seed}

\item{feature_data}{feature annotation data frame}

\item{annot_row}{select colnames of feature_data}

\item{smooth_n}{smooth the expression over smooth_n cells}

\item{do_scale}{scale data}

\item{smooth_lowess}{smooth using lowess regression}

\item{lowess_span}{lowess span}

\item{cap_max}{capping the maximum of data}

\item{cap_min}{capping the minimum of data}

\item{annot_colors}{a list of annotation colors named with annotation categories}

\item{rownames_to_show}{select rownames to show to the right}

\item{...}{other params passed to pheatmap}
}
\value{
a pheatmap
}
\description{
plot heatmap using pheatmap
}
\examples{

}
