% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_intersectHeatmap}
\alias{pl_intersectHeatmap}
\title{upper triangle heatmap of pairwise intersectin matrix}
\usage{
pl_intersectHeatmap(
  mat,
  tri_fun = upper.tri,
  hide_diag = T,
  proportion_or_roe = c("roe", "proportion"),
  dist = "correlation",
  color = colorRampPalette(c("grey90", RColorBrewer::brewer.pal(n = 7, name =
    "YlOrRd")))(100),
  clustering_method = "ward.D",
  treeheight_row = 25,
  treeheight_col = 25,
  border_color = "white",
  ...
)
}
\arguments{
\item{mat}{matrix of 0 or 1 and 1 means the column has the rowname}

\item{tri_fun}{upper.tri or lower.tri}

\item{hide_diag}{hide diag color}

\item{proportion_or_roe}{calculate scaled item}

\item{dist}{distance method}

\item{color}{vector of colors used in heatmap.}

\item{clustering_method}{clustering method used. Accepts the same values as 
\code{\link{hclust}}.}

\item{treeheight_row}{the height of a tree for rows, if these are clustered. 
Default value 50 points.}

\item{treeheight_col}{the height of a tree for columns, if these are clustered. 
Default value 50 points.}

\item{border_color}{color of cell borders on heatmap, use NA if no border should be 
drawn.}

\item{...}{other params passed to pheatmap}
}
\value{
pheatmap
}
\description{
upper triangle heatmap of pairwise intersectin matrix
}
\examples{

}
