% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_rastrFeaturePlot}
\alias{pl_rastrFeaturePlot}
\title{rastr version of partial FeaturePlot}
\usage{
pl_rastrFeaturePlot(
  object,
  features,
  coord_fixed_ratio = 1,
  ncol = NULL,
  align = c("hv"),
  cols = c("grey", "yellow", "red"),
  pt.size = 1,
  order = T,
  rasterise = T,
  reduction = "umap",
  dims = c(1, 2)
)
}
\arguments{
\item{object}{seurat object}

\item{features}{genes}

\item{coord_fixed_ratio}{coord_fixed}

\item{ncol}{ncol of cowplot::plot_grid}

\item{align}{align of cowplot::plot_grid}

\item{cols}{colors of gradientn}

\item{pt.size}{point size}

\item{order}{sort orders of cells}

\item{reduction}{dimension reduction}

\item{dims}{dimensions}
}
\value{
ggplot
}
\description{
rastr version of partial FeaturePlot
}
\examples{

}
