% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_segmentPlot}
\alias{pl_segmentPlot}
\title{Plot the expression of one gene using bar/vline, or so called barplot}
\usage{
pl_segmentPlot(x, cluster)
}
\arguments{
\item{x}{numeric vector, such as expression of one gene in all cells.}

\item{cluster}{character vector of cell annotation, such as cell clusters.}
}
\value{
a ggplot object
}
\description{
Plot the expression of one gene using bar/vline, or so called barplot
}
\examples{

}
