% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_slingshot}
\alias{pl_slingshot}
\title{DimPlot and Slingshot curves}
\usage{
pl_slingshot(object, ggplot, color_by = "ss_lineage", pt.size = 1)
}
\arguments{
\item{object}{Seurat3 object}

\item{ggplot}{a ggplot object as background}

\item{color_by}{slingshot curve color_by}
}
\value{
a ggplot object
}
\description{
DimPlot and Slingshot curves
}
\examples{

}
