% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_splitDimplot}
\alias{pl_splitDimplot}
\title{plot multiple DimPlot for each batch}
\usage{
pl_splitDimplot(
  object,
  reduction = "umap",
  split_by = "Dataset",
  groups_label = c("seurat_clusters"),
  groups = c("Batch", "Phase"),
  coord_fixed_ratio = 1,
  pt.size = 1,
  colors = scanpy_colors$default_64
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{umap/tsne/pca}

\item{split_by}{split dimplot by this metadata}

\item{groups_label}{plot using DimPlot with label = TRUE}

\item{groups}{plot using pl_dimplot}

\item{coord_fixed_ratio}{only valid for groups}

\item{pt.size}{point size}

\item{colors}{point color}
}
\value{
multiple plots
}
\description{
This may be useful for displaying multiple datasets
}
\examples{

}
