% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_stackedViolinPlot}
\alias{pl_stackedViolinPlot}
\title{stacked Violin Plots in one plot}
\usage{
pl_stackedViolinPlot(
  data,
  features,
  group_by,
  groups = NULL,
  fill_colors = NULL,
  ncol = 1,
  facet_scales = "free_y"
)
}
\arguments{
\item{data}{expression matrix of feature by cell, ususally normalized (in log scale) but not feature scaled.}

\item{features}{features to be visualized}

\item{group_by}{a vector including cell annotations}

\item{groups}{only these selected groups to be visualized}

\item{fill_colors}{violion fill colors}

\item{ncol}{number of columns to plot}

\item{facet_scales}{scales of facet_wrap}
}
\value{
a ggplot object
}
\description{
stacked Violin Plots in one plot
}
