% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_tableHeatmap}
\alias{pl_tableHeatmap}
\title{Heatmap of a 2D-table}
\usage{
pl_tableHeatmap(
  tab,
  palette = "YlGn",
  n = 5,
  min = 0,
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{tab}{a table object}

\item{palette}{palette name in RColorBrewer or a vector of gradient colors}

\item{n}{length of gradient colors}

\item{min}{less than min will be setted to NA}

\item{title}{plot title}

\item{xlab}{plot x lab}

\item{ylab}{plot y lab}
}
\value{
a ggplot object
}
\description{
Heatmap of a 2D-table
}
\examples{

}
