% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_upset_cluster_regulon}
\alias{pl_upset_cluster_regulon}
\title{upset plot wrapper for intersection of regulons}
\usage{
pl_upset_cluster_regulon(
  regulon_cluster_mat,
  clusters = NULL,
  degs_table = NULL,
  nintersects = NA,
  colors_clusters = NULL,
  color_queries = "tomato",
  ...
)
}
\arguments{
\item{regulon_cluster_mat}{a matrix/data.frame of 0 or 1, rownames are regulons, colnames are clusters}

\item{clusters}{clusters to be plotted. set NULL for all colnames of `regulon_cluster_mat`}

\item{degs_table}{table of degs derived from FindAllMarkers. columns of `gene` and `cluster` is required}

\item{nintersects}{set NA for all to be shown}

\item{colors_clusters}{colors for clusters}

\item{color_queries}{color for queried bar}

\item{...}{other params passed to UpsetR::upset}
}
\value{
plot
}
\description{
upset plot wrapper for intersection of regulons
}
\examples{

}
