% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pmultinom_sample}
\alias{pmultinom_sample}
\title{calculate the probability of collecting all kinds of coupons when collecting n coupons.}
\usage{
pmultinom_sample(
  prob,
  k = length(prob),
  n.max = 100,
  B = 1000,
  seed = 66,
  replace = TRUE
)
}
\arguments{
\item{prob}{probabilitys for each kind of coupon}

\item{k}{default length of prob}

\item{n.max}{collected 1:n.max coupons}

\item{B}{an integer specifying the number of replicates used in the Monte Carlo test.}

\item{seed}{for reproducibility}

\item{replace}{sample wi/o replacing}
}
\value{
a vector of probabilities for 1:n.max cllections
}
\description{
calculate the probability of collecting all kinds of coupons when collecting n coupons.
}
\examples{
# Given that 6 kinds of coupons, of which each's probability is 1/6,
# Let's see the probability of collection all kinds of coupons for having n coupons.

pmultinom_sample(rep(1,6)/6)
}
