% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_addMitoRatio}
\alias{pp_addMitoRatio}
\title{Add Mitochondrial genes ratio}
\usage{
pp_addMitoRatio(object, nCount = "nCount_RNA", mito_genes = NULL)
}
\arguments{
\item{object}{seurat object}

\item{nCount}{colname of meta.data for UMI count}

\item{mito_genes}{genes to be used as Mitochondrial genes. default: `grepl("^mt-", rownames(object), ignore.case = T)`}
}
\value{
updated seurat object
}
\description{
Add Mitochondrial genes ratio
}
\examples{

}
