% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_annotateHashtagSeurat}
\alias{pp_annotateHashtagSeurat}
\title{Annotate seurat data derived by pp_hashtagDemultiplex function}
\usage{
pp_annotateHashtagSeurat(files, metadata = NULL)
}
\arguments{
\item{files}{seurat Rdata files produced by pp_hashtagDemultiplex}

\item{metadata}{annotation information for each file/seurat_object.
e.g. list(c(Batch = "b1"), c(Batch = "b2"), NULL, c(Batch = "b4"))}
}
\value{
a list of seurat objects
}
\description{
Annotate seurat data derived by pp_hashtagDemultiplex function
}
\examples{

}
