% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_centerData}
\alias{pp_centerData}
\title{Scale data in each group}
\usage{
pp_centerData(
  data,
  group_by = NULL,
  by_log = F,
  log_base = base::exp(1),
  center = T,
  scale = T
)
}
\arguments{
\item{data}{expression matrix of feature by cell, ususally normalized (in log scale) but not feature scaled.}

\item{group_by}{a vector including cell annotations}

\item{by_log}{calculate in log scale or in normalized scale; If TRUE log_base will be ignored.}

\item{log_base}{base::exp(1) or 2 or other else, depends on your normalization method.}

\item{center}{a logical value used in base::scale}

\item{scale}{a logical value used in base::scale}
}
\value{
scaled data
}
\description{
Scale data in each group
}
\examples{

}
