% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_fastMNN}
\alias{pp_fastMNN}
\title{A wrapper function to quickly run fastMNN using SeuratWrappers::RunFastMNN}
\usage{
pp_fastMNN(
  object_list,
  genes_used = NULL,
  batch_by = "Batch",
  scale.factor = 1e+05,
  s.features = cc.genes$s.genes,
  g2m.features = cc.genes$g2m.gene,
  assay = c("RNA", "SCT"),
  nfeatures = 2000,
  integration_features = NULL,
  remove_ccgenes_by_cor = T,
  cc_cor_th = 0.3,
  mnn_d = 20,
  seed = 666,
  skip_tsne = F,
  resolution = 0.5,
  save_object_prefix = NULL,
  save_degs_dir = "tables/",
  logfc.threshold = log(2),
  min.pct = 0.25,
  save_rdata_dir = "seuratData/",
  save_add.sys.time = F,
  ...
)
}
\arguments{
\item{object_list}{a list of Seurat objects}

\item{genes_used}{Set NULL to use the intersection of all Seurat objects.}

\item{batch_by}{Attribute for splitting. Default is "Batch". (SplitObject)}

\item{scale.factor}{Sets the scale factor for cell-level normalization. Calculated in RNA assay. (NormalizeData)}

\item{s.features}{A vector of features associated with S phase. Calculated in RNA assay. (CellCycleScoring)}

\item{g2m.features}{A vector of features associated with G2M phase. Calculated in RNA assay. (CellCycleScoring)}

\item{assay}{which assay to be used for integration. default RNA.}

\item{nfeatures}{Number of features to return (SelectIntegrationFeatures)}

\item{integration_features}{if NULL use nfeatures, else use these feature for fastMNN}

\item{remove_ccgenes_by_cor}{whether to exclude cell cycle-related genes using correlation method.}

\item{cc_cor_th}{feature genes with correlation coefficient larger than this threshold will be removed.}

\item{mnn_d}{Number of dimensions to use for dimensionality reduction in multiBatchPCA. (batchelor::fastMNN)}

\item{seed}{random seed for fastMNN}

\item{skip_tsne}{you may set TRUE to skip RunTSNE step for large dataset.}

\item{resolution}{Value of the resolution parameter (FindClusters)}

\item{save_object_prefix}{Set NULL to skip save files and FindAllMarker steps. Or a character setting the object prefix, e.g. AAA will get AAA_mnn and AAA_degs being returned.}

\item{save_degs_dir}{{save_object_prefix}_degs.seurat_clusters.{sys.time}.xls will be written to here.}

\item{logfc.threshold}{Limit testing to genes which show at least X-fold difference (log-scale) between the two groups of cells. (FindAllMarkers)}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. (FindAllMarkers)}

\item{save_rdata_dir}{{save_object_prefix}_mnn.seurat.{sys.time}.Rdata will be written to here.}

\item{save_add.sys.time}{whether to add sys.time to file names.}

\item{...}{Extra parameters passed to batchelor::fastMNN.}
}
\value{
seurat object if save_object_prefix is NULL, else saved files.
}
\description{
Run NormalizeData, CellCycleScoring, SelectIntegrationFeatures, RunFastMNN, RunUMAP/TSNE, FindClusters and FindMarkers.
And save object_mnn and object_degs to xls and Rdata files.
}
\examples{

}
