% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_findVariableFeatures}
\alias{pp_findVariableFeatures}
\title{Find variable features across batches}
\usage{
pp_findVariableFeatures(
  object,
  split_by = "ident",
  groups = NULL,
  nfeatures = 2000,
  assay = "RNA",
  log_var = F,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{split_by}{a colname of meta.data}

\item{groups}{only some groups of split_by are considered.}

\item{nfeatures}{Number of features to select as top variable features; only used when selection.method is set to 'dispersion' or 'vst'}

\item{...}{Arguments passed to Seurat::FindVariableFeatures}
}
\value{
A vector of variable features
}
\description{
Identifies Highly variable features (HVFs) for each batch, then combine them into a single vector of HVFs by count.
}
\examples{

}
