% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_hashtagDemultiplex}
\alias{pp_hashtagDemultiplex}
\title{render rmarkdown file for de-multiplex hash-tag data of h5file generated by cellranger}
\usage{
pp_hashtagDemultiplex(
  h5file,
  seuratfile,
  sample,
  hashtag = stringr::str_glue("B025{1:6}_anti-human_Hashtag_{1:6}"),
  output_format = "html_document",
  output_file = NULL,
  output_dir = "./"
)
}
\arguments{
\item{h5file}{.h5 file generated by cellranger}

\item{seuratfile}{save seurat obejct into this file, eg. 'hashtag.seurat.Rdata'}

\item{sample}{real sample names represented by hash-tag, must be corresponding to hashtag each by each.}

\item{hashtag}{hash-tag names used in the feature-ref file of your cellranger pipeline}

\item{output_format}{The R Markdown output format to convert to. The option
\code{"all"} will render all formats defined within the file. The option can
be the name of a format (e.g. \code{"html_document"}) and that will render
the document to that single format. One can also use a vector of format
names to render to multiple formats. Alternatively, you can pass an output
format object (e.g. \code{html_document()}). If using \code{NULL} then the
output format is the first one defined in the YAML frontmatter in the input
file (this defaults to HTML if no format is specified there).}

\item{output_file}{The name of the output file. If using \code{NULL} then the
output filename will be based on filename for the input file. If a filename
is provided, a path to the output file can also be provided. Note that the
\code{output_dir} option allows for specifying the output file path as well,
however, if also specifying the path, the directory must exist. If
\code{output_file} is specified but does not have a file extension, an
extension will be automatically added according to the output format. To
avoid the automatic file extension, put the \code{output_file} value in
\code{\link{I}()}, e.g., \code{I('my-output')}.}

\item{output_dir}{The output directory for the rendered \code{output_file}.
This allows for a choice of an alternate directory to which the output file
should be written (the default output directory of that of the input file).
If a path is provided with a filename in \code{output_file} the directory
specified here will take precedence. Please note that any directory path
provided will create any necessary directories if they do not exist.}
}
\value{
derived html file
}
\description{
render rmarkdown file for de-multiplex hash-tag data of h5file generated by cellranger
}
\details{
Given that this package is designed for human, mouse and macaca,
when Mitochondria genes did not start with "mt-" (case was ignored. This situation should only occur in Macaca),
Ensembl gene id (already stored in this package) of Macaca mulatta will be used.
}
\examples{

}
\seealso{
\code{\link{rmarkdown::render}}
}
