% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_preprocess}
\alias{pp_preprocess}
\title{A wrapper function to quickly run standard processing pipeline}
\usage{
pp_preprocess(
  object,
  scale.factor = 10000,
  s.features = cc.genes$s.genes,
  g2m.features = cc.genes$g2m.gene,
  do.regress.cc = T,
  do.scale = T,
  pp = c("SCT", "RNA", "centerData"),
  center_by = NULL,
  nfeatures = 2000,
  remove_ccgenes_by_cor = F,
  ndim = 20,
  skip_tsne = F,
  resolution = 0.5
)
}
\arguments{
\item{object}{a Seurat object}

\item{scale.factor}{Sets the scale factor for cell-level normalization. (NormalizeData)}

\item{s.features}{A vector of features associated with S phase (CellCycleScoring)}

\item{g2m.features}{A vector of features associated with G2M phase (CellCycleScoring)}

\item{do.regress.cc}{regress out cell cycle scores (ScaleData)}

\item{do.scale}{Whether to scale the data. (ScaleData)}

\item{pp}{do SCTransform or standard RNA process}

\item{center_by}{set a meta name when pp == "centerData" (pp_centerData)}

\item{nfeatures}{Number of features to return (SCTransform/FindVariableFeatures)}

\item{remove_ccgenes_by_cor}{Only used in RNA or centerData pipeline. Remove ccgenes generated by cor analysis (cor > 0.3) from hvgs.}

\item{ndim}{Number of dimensions to use for dimensionality reduction and clustering (RunUMAP/TSNE/FindClusters)}

\item{skip_tsne}{you may set TRUE to skip RunTSNE step for large dataset.}

\item{resolution}{Value of the resolution parameter (FindClusters)}
}
\value{
a Seurat object
}
\description{
Run NormalizeData, CellCycleScoring, FindVariableFeatures, ScaleData, RunUMAP/TSNE and FindClusters.
}
\examples{

}
