% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_removeCCgenesByCor}
\alias{pp_removeCCgenesByCor}
\title{Remove ccgenes generated by cor analysis (cor > 0.3) from hvgs.}
\usage{
pp_removeCCgenesByCor(object, features = NULL, cor_th = 0.3)
}
\arguments{
\item{object}{seurat object}

\item{features}{set NULL to use VariableFeatures}

\item{cor_th}{threshold above which presumed to be cc genes.}
}
\value{
seurat object
}
\description{
Remove ccgenes generated by cor analysis (cor > 0.3) from hvgs.
}
\examples{

}
