% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_runHarmony}
\alias{pp_runHarmony}
\title{A Wrapper of harmony::RunHarmony}
\usage{
pp_runHarmony(object, dims.use, group.by.vars, ...)
}
\arguments{
\item{object}{Seurat object}

\item{dims.use}{Which PCA dimensions to use for Harmony. By default, use all}

\item{group.by.vars}{Which variable(s) to remove (character vector).}

\item{...}{other parameters to be used in harmony::HarmonyMatrix}
}
\value{
Seurat (version 3) object.
}
\description{
RunHarmony did not support dims.use actually, so I wrote this wrapper function to support that.
}
