% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp_transform2dDim}
\alias{pp_transform2dDim}
\title{transform reduction coordinates}
\usage{
pp_transform2dDim(
  object,
  reduction = "umap",
  transform_matrix = crosslink::transform_matrix_affine(type = "reflect", theta = 0)
)
}
\arguments{
\item{object}{seurat3 object}

\item{reduction}{umap or tsne or something}

\item{transform_matrix}{3 x 3 matrix derived from crosslink::transform_matrix_affine}
}
\value{
seurat3 object
}
\description{
transform reduction coordinates
}
\examples{

}
