% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{presuf_file_name}
\alias{presuf_file_name}
\title{Add prefix and/or suffix for file name}
\usage{
presuf_file_name(file_name, prefix = NULL, suffix = NULL, sep = ".", index = 1)
}
\arguments{
\item{file_name}{file name to be added}

\item{prefix}{prefix character}

\item{suffix}{suffix character}

\item{sep}{split file name by sep into multiple parts}

\item{index}{add prefix to the index part. if <= 0, reverse order is used, then 0 is the last one and -1 is the last but one.}
}
\value{
new file name
}
\description{
Add prefix and/or suffix for file name
}
\examples{
presuf_file_name("./database/msigdb.gmt", "homo_")
# "./database/homo_msigdb.gmt"

}
