% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.gmt}
\alias{read.gmt}
\title{Read gmt file}
\usage{
read.gmt(gmtfile = NULL)
}
\arguments{
\item{gmtfile}{file name}
}
\value{
a list of genesets, each with both gene list and description list
}
\description{
Read gmt file
}
\examples{

}
