% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readDataFrameFromClipboard}
\alias{readDataFrameFromClipboard}
\title{read table from clipboard}
\usage{
readDataFrameFromClipboard(header = T, sep = "\\t", check.names = F, ...)
}
\arguments{
\item{header}{a logical value indicating whether the file contains the
    names of the variables as its first line.  If missing, the value is
    determined from the file format: \code{header} is set to \code{TRUE}
    if and only if the first row contains one fewer field than the
    number of columns.}

\item{sep}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}

\item{check.names}{logical.  If \code{TRUE} then the names of the
    variables in the data frame are checked to ensure that they are
    syntactically valid variable names.  If necessary they are adjusted
    (by \code{\link{make.names}}) so that they are, and also to ensure
    that there are no duplicates.}

\item{...}{other params passed to \code{\link{utils::read.table}}}
}
\value{
a data.frame
}
\description{
read table from clipboard
}
\examples{

}
