% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readRda}
\alias{readRda}
\title{load saved RData}
\usage{
readRda(file, verbose = F)
}
\arguments{
\item{file}{a (readable binary-mode) connection or a character string giving the name of the file to load (when tilde expansion is done).}

\item{verbose}{should item names be printed during loading?}
}
\value{
A list of objects created.
}
\description{
A wrapper function of base::load
This is usefull when the names of object conflict with variables in .GlobalEnv
}
\examples{

}
