% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regressionSmoothen}
\alias{regressionSmoothen}
\title{Regression based smoothened line prediction}
\usage{
regressionSmoothen(
  x,
  y,
  method = stats::loess,
  formula = y ~ x,
  span = 0.75,
  ...
)
}
\arguments{
\item{x}{independent variable, numeric vector}

\item{y}{dependent variable, numeric vector}

\item{method}{Smoothing function, see ggplot2::geom_smooth}

\item{formula}{Formula to use in smoothing function, eg. y ~ x, y ~ poly(x, 2), y ~ log(x).}

\item{span}{Controls the amount of smoothing for the default loess smoother.}

\item{...}{other parameters used in `method` function}
}
\value{
data.frame of fit and se.fit?
}
\description{
Regression based smoothened line prediction
}
\examples{

}
