% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieveKEGG}
\alias{retrieveKEGG}
\title{download updated KEGG Pathways}
\usage{
retrieveKEGG(org = c("hsa", "mmu"), symbol = TRUE, gmt = TRUE)
}
\arguments{
\item{org}{supported organism, can be search using clusterProfiler::search_kegg_organism function}

\item{symbol}{return gene symbol}

\item{gmt}{return as gmt format. Use write.gmt to export to file.}
}
\value{
kegg terms
}
\description{
download updated KEGG Pathways
}
\examples{

}
