% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{revLevels}
\alias{revLevels}
\title{revert vector's levels}
\usage{
revLevels(x, levels = NULL)
}
\arguments{
\item{x}{a vector}

\item{levels}{relevel x using `levels`}
}
\value{
updated x with levels reverted
}
\description{
revert vector's levels
}
\examples{

}
