% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RobustRescale}
\alias{RobustRescale}
\title{similar to scales::rescale but use quantile 95}
\usage{
RobustRescale(x, probs = c(0.05, 0.95), max = 10, min = 0, ground = NULL)
}
\arguments{
\item{x}{a vector of numeric}

\item{probs}{get the actual_min and actual_max. this may remove the outliers effect.}

\item{max}{rescale actual_max to this max value}

\item{min}{rescale actual_min to this min value}

\item{ground}{for gene expression , to set 0 as the actual_min}
}
\value{
a vector
}
\description{
similar to scales::rescale but use quantile 95
}
\examples{

}
