% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowSmooth}
\alias{rowSmooth}
\title{smooth each row of matrix}
\usage{
rowSmooth(mat, n = NULL, proportion = 0.2, do.scale = F)
}
\arguments{
\item{mat}{matrix}

\item{n}{smooth over n columns}

\item{proportion}{smooth over proportion*ncol(mat)}

\item{do.scale}{scale mat after being smoothed}
}
\value{
matrix
}
\description{
smooth each row of matrix
}
\examples{

}
