% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_fun_n_times}
\alias{run_fun_n_times}
\title{Run function n times with output as input}
\usage{
run_fun_n_times(fun, x, times = 1, ...)
}
\arguments{
\item{fun}{function}

\item{x}{input also output}

\item{times}{times}

\item{...}{other params}
}
\value{
x
}
\description{
Run function n times with output as input
}
\examples{

}
