% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_cluster_cells}
\alias{sample_cluster_cells}
\title{sampling cells by clusters}
\usage{
sample_cluster_cells(cells, clusters, n = 60, seed = 666, unlist = F)
}
\arguments{
\item{cells}{vector}

\item{clusters}{vector}

\item{n}{sampling number of cells}

\item{seed}{random seed}

\item{unlist}{return list or not}
}
\value{
subset of cells
}
\description{
sampling cells by clusters
}
\examples{

}
