% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{scanpy_colors}
\alias{scanpy_colors}
\title{discrete colors imported from scanpy.pl.palettes}
\format{
A list with 4 different palettes of 10, 20, 26, 64 colors.
}
\source{
\url{https://github.com/theislab/scanpy/issues/387},
\url{https://github.com/vega/vega/wiki/Scales#scale-range-literals},
\url{https://graphicdesign.stackexchange.com/questions/3682/where-can-i-find-a-large-palette-set-of-contrasting-colors-for-coloring-many-d},
\url{http://godsnotwheregodsnot.blogspot.de/2012/09/color-distribution-methodology.html}
}
\usage{
scanpy_colors
}
\description{
A list containing the colors used in scanpy
}
\keyword{datasets}
