% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SetDefaultAes}
\alias{SetDefaultAes}
\title{update data.frame with new name added}
\usage{
SetDefaultAes(df, x = NULL, default = 1, name = "shape")
}
\arguments{
\item{df}{data.frame}

\item{x}{NULL or a value of length one}

\item{default}{use this value is x is NULL}

\item{name}{the name to be added to df}
}
\value{
a new df with name added
}
\description{
update data.frame with new name added
}
\examples{

}
