% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setDFMode}
\alias{setDFMode}
\title{change data.frame storage mode}
\usage{
setDFMode(x, mode = c("character", "numeric"))
}
\arguments{
\item{x}{a data.frame}

\item{mode}{a character}
}
\value{
updated data.frame
}
\description{
easy to convert numeric to character, or vice versa
}
\examples{

}
