% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setLevels}
\alias{setLevels}
\title{set levels for a given column}
\usage{
setLevels(df, x, levels)
}
\arguments{
\item{df}{data.frame}

\item{x}{a colname of \code{df}}

\item{levels}{levels for \code{df[[x]]}}
}
\value{
updated data.frame
}
\description{
set levels for a given column
}
\examples{

}
