% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setSummary}
\alias{setSummary}
\title{set calculations}
\usage{
setSummary(x, y)
}
\arguments{
\item{x}{set A}

\item{y}{set B}
}
\value{
a list of diffs, intersect and union
}
\description{
set calculations
}
\examples{

}
