% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SignCorrectionByCor}
\alias{SignCorrectionByCor}
\title{Correct sign by correlation}
\usage{
SignCorrectionByCor(x, y)
}
\arguments{
\item{x}{a vector to be corrected}

\item{y}{a vector as reference with equal length of x}
}
\value{
corrected x
}
\description{
Correct sign by correlation
}
\examples{

}
