% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to.R
\name{str_to_char}
\alias{str_to_char}
\title{convert string into a vector of each char}
\usage{
str_to_char(x, unique = F, sort = F, ...)
}
\arguments{
\item{x}{a string or a vector of strings. Note if a vector supplied, it will be collapsed. If you wanna multiple vectors of char returned, you can use sapply() to wrapper the function.}

\item{unique}{If `TRUE` chars should be uniqued, otherwise as it is.}

\item{sort}{If `TRUE` chars should be sorted by alphabetic, otherwise as it is.}

\item{...}{more parameters to be used in sort(). such as order decreasing = T if you like.}
}
\value{
a vector of chars
}
\description{
convert string into a vector of each char
}
\examples{
str_to_char(x = c("Here it is"), unique = F, sort = F)
str_to_char(x = c("Yes", "or", "Not"), unique = F, sort = F)
sapply(c("Yes", "or", "Not"), str_to_char, sort = F, unique = F)
}
