% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SubsetDataAndGroup}
\alias{SubsetDataAndGroup}
\title{Subset Data and Group}
\usage{
SubsetDataAndGroup(data, group_by, groups = NULL, droplevels = T)
}
\arguments{
\item{data}{expression matrix of feature by cell. Sparse matrix will be densed, so be caution of your memory.}

\item{group_by}{a vector including cell annotations}

\item{groups}{subset these selected groups. set NULL to use all.}

\item{droplevels}{drop the unused levels in group_by}
}
\value{
a list of data and group_by (factor)
}
\description{
Subset Data and Group
}
\examples{

}
