% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SummariseExpr}
\alias{SummariseExpr}
\title{Do summarise (e.g. mean) for a given vector}
\usage{
SummariseExpr(
  x,
  method = base::mean,
  by_log = F,
  log_base = base::exp(1),
  exclude_zero = F,
  outlier_cutoff = 1
)
}
\arguments{
\item{x}{numeric values in log scale}

\item{method}{base::mean or any other function returned a value}

\item{by_log}{calculate by log scale ; If TRUE log_base will be ignored.}

\item{log_base}{base::exp(1) or 2 or other else, depends on your normalization method.}

\item{exclude_zero}{exclude zeros when calculate average/median feature value. Note exclude_zero first, outlier_cutoff second.}

\item{outlier_cutoff}{sometimes outliers (several extremely high cells) should be excluded when do summarise. Set 0.99 to exclude top 1 percent cells. (default: 1)}
}
\value{
a numeric value
}
\description{
Do summarise (e.g. mean) for a given vector
}
\examples{

}
