% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_AddABCScore}
\alias{tl_AddABCScore}
\title{Quick add module scores for markers from Atlas of Human Blood Cells}
\usage{
tl_AddABCScore(object, topn = 50, ...)
}
\arguments{
\item{object}{Seurat3 object}

\item{topn}{top n marker genes for each cluster}

\item{...}{other parameters passed to `Seurat::AddModuleScore`}
}
\value{
Seurat Object
}
\description{
Quick add module scores for markers from Atlas of Human Blood Cells
}
\details{
http://scrna.sklehabc.com/
Marker Gene: Information of 43 transcriptional cell clusters and signature genes with AUC score.
National Science Review, Volume 8, Issue 3, March 2021, nwaa180, https://doi.org/10.1093/nsr/nwaa180
}
\examples{

}
