% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_AddReduction}
\alias{tl_AddReduction}
\title{Add new Reduction to Seurat object}
\usage{
tl_AddReduction(object, reduction, dim.1, dim.2, ...)
}
\arguments{
\item{object}{Seurat v3 object}

\item{reduction}{reduction name}

\item{dim.1}{vector of 1st dimension}

\item{dim.2}{vector of 2nd dimension}

\item{...}{other params passed to CreateDimReducObject()}
}
\value{
updated Seurat object
}
\description{
Add new Reduction to Seurat object
}
\examples{

}
