% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_calcFeatureScore}
\alias{tl_calcFeatureScore}
\title{calculate average/median rescale expression of features (eg. genes) for each cell}
\usage{
tl_calcFeatureScore(
  data,
  features,
  rescale_max = 10,
  rescale_min = 0,
  method = base::mean,
  by_log = T,
  log_base = base::exp(1),
  exclude_zero = F,
  outlier_cutoff = 1,
  cap_value = NULL,
  gene_force = F,
  rescale_again = T
)
}
\arguments{
\item{data}{expression matrix of feature by cell, ususally normalized (in log scale) but not feature scaled. A matrix is better than data.frame in case of duplicated features.}

\item{features}{features to be calculated}

\item{rescale_max}{rescale data}

\item{rescale_min}{rescale data}

\item{method}{using base::mean or stats::median or your custom function for each feature vector;}

\item{by_log}{calculate by log scale ; If TRUE log_base will be ignored.}

\item{log_base}{base::exp(1) or 2 or other else, depends on your normalization method.}

\item{exclude_zero}{exclude zeros when calculate average/median feature value. Note exclude_zero first, outlier_cutoff second.}

\item{outlier_cutoff}{sometimes outliers (several extremely high cells) should be excluded when do summarise. Set 0.99 to exclude top 1 percent cells. (default: 1)}

\item{cap_value}{the max value to show in dotplot. Any value larger than it will be capped to this value. (default: NA)}

\item{gene_force}{Force gene using zeros, if it is not in data.}

\item{rescale_again}{do rescale again after summarisie}
}
\value{
a vector of scores
}
\description{
calculate average/median rescale expression of features (eg. genes) for each cell
}
\examples{

}
