% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_calcSimilarity}
\alias{tl_calcSimilarity}
\title{calculate inter-similarity or intra-similarity of clusters}
\source{
Fig. 4A at https://doi.org/10.1016/j.cell.2019.11.010
}
\usage{
tl_calcSimilarity(
  data,
  group_by,
  groups = NULL,
  subset_by = NULL,
  subsets = NULL,
  dist_fun = stats::dist,
  metric = "euclidean",
  transform_to_similarity = T,
  transform_method = function(x) 1(1 + x),
  ...
)
}
\arguments{
\item{data}{feature by cell matrix}

\item{group_by}{vector of cell annotation}

\item{groups}{only consider subset cells belonging to these groups}

\item{subset_by}{if setted, calc intra-/inter-similarity of groups restricted to each subset}

\item{subsets}{only consider cells belonging to these subsets}

\item{dist_fun}{function to calculate distance/similarity. see Details.}

\item{metric}{metric passed to dist_fun}

\item{transform_method}{dist to similarity function, see Details}

\item{...}{ohter params passed to dist_fun}
}
\value{
a slimilarity matrix, usually 0-1
}
\description{
calculate inter-similarity or intra-similarity of clusters
}
\details{
dist_fun:
stats::dist,
ClassDiscovery::distanceMatrix (pearson, sqrt pearson, spearman, absolute pearson and so on),
philentropy::distance (46 different distances/similarities),

transform_method:
1) 1/(1+x)
2) 1-x/max(x)
3) sqrt(1-x/max(x))
4) 1-x
5) 1-abs(x)
any meaningful function
}
\examples{

}
