% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_CellCycleAssign}
\alias{tl_CellCycleAssign}
\title{Reassign cell cycle pahse and set cc reduction}
\usage{
tl_CellCycleAssign(
  object,
  phase.name = "Phase4",
  s.th = 0,
  g2m.th = 0,
  set.reduction = "cc",
  slope = 1,
  s.scores = "S.Score",
  g2m.scores = "G2M.Score"
)
}
\arguments{
\item{object}{Seurat v3 object}

\item{phase.name}{name added to meta.data}

\item{s.th, g2m.th, slope}{threshold of S.Score and G2M.Score used to reassign phase, and slope is used for partition S and G2/M}

\item{set.reduction}{set NULL if do not want to add cc reduction, Or set a character value for reduction name}

\item{s.scores}{vector of S.Scores}

\item{g2m.scores}{vector of G2M.Scores}
}
\value{
updated Seurat object
}
\description{
Reassign cell cycle pahse and set cc reduction
}
\examples{

}
