% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_crossTable}
\alias{tl_crossTable}
\title{Plot proportions}
\usage{
tl_crossTable(
  tab,
  proportion_by = c("column", "row"),
  ret_value = F,
  color = colorRampPalette(c("white", "orange", "red"))(20),
  cluster_rows = F,
  cluster_cols = F,
  display_numbers = T,
  number_format = "\%1.1f\%\%",
  ...
)
}
\arguments{
\item{tab}{table}

\item{proportion_by}{row or column}

\item{ret_value}{return proportions or heatmap}

\item{...}{other params used for pheatmap}
}
\value{
proportions or heatmap
}
\description{
Plot proportions
}
\examples{

}
