% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_geneExpressed}
\alias{tl_geneExpressed}
\title{categorize gene as expressed or not based on detection rate and non-zero median expression level}
\usage{
tl_geneExpressed(data, group_by = NULL)
}
\arguments{
\item{data}{feature by cell matrix}

\item{group_by}{cell annotation vector}
}
\value{
data.frame of detection, nzmedian, expressed, cluster
}
\description{
categorize gene as expressed or not based on detection rate and non-zero median expression level
}
\examples{

}
