% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_GeneSetScore}
\alias{tl_GeneSetScore}
\title{Gene Set Scores}
\usage{
tl_GeneSetScore(
  object,
  genesets,
  method = c("pca", "moduleScore", "gsva", "ssgsea", "zscore", "plage"),
  ...
)
}
\arguments{
\item{object}{Seurat3 object}

\item{genesets}{a list of one or more sets of features}

\item{method}{supported methods. It turns the sparse matrix into dense matrix for methods in GSVA.}

\item{...}{other params passed to corresponding method}
}
\value{
a list of vectors of score values
}
\description{
Gene Set Scores
}
\examples{

}
