% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_go}
\alias{tl_go}
\title{A wrapper function for clusterprofiler GO enrichment analysis (enrichGO)}
\usage{
tl_go(
  gene,
  simplify = F,
  OrgDb = "org.Hs.eg.db",
  keyType = "SYMBOL",
  ont = "BP",
  ...
)
}
\arguments{
\item{gene}{a vector of focused genes}

\item{simplify}{do simplify or not}

\item{OrgDb}{org.Hs.eg.db or org.Mm.eg.db, or else}

\item{keyType}{SYMBOL or ENTREZID, or else}

\item{ont}{BP, MF, CC}

\item{...}{other parameters for enrichGO}
}
\value{
a list of GO and simplified GO results
}
\description{
A wrapper function for clusterprofiler GO enrichment analysis (enrichGO)
}
\examples{

}
