% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_knnClassify}
\alias{tl_knnClassify}
\title{Using class::knn method in Dimension Reduction space to classify cells with referenced clusters}
\usage{
tl_knnClassify(
  object,
  reference_group_by,
  predict_group_by = "knnClassify",
  reduction = "mnn",
  dims = NULL,
  k = 20,
  reference_cells = NULL,
  predict_cells = NULL
)
}
\arguments{
\item{object}{seurat3 object}

\item{reference_group_by}{a meta.data colname}

\item{reduction}{dimension reduction}

\item{dims}{use all dims if NULL, or specify it}

\item{k}{number of kNN used to predict cluster}

\item{reference_cells}{cells used as reference}

\item{predict_cells}{cells to be predicted/classified}
}
\value{
seurat object
}
\description{
Using class::knn method in Dimension Reduction space to classify cells with referenced clusters
}
\examples{

}
