% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_OrderGene}
\alias{tl_OrderGene}
\title{Prepare genelist for GSEA}
\usage{
tl_OrderGene(
  data,
  group.by,
  group.1,
  group.2 = NULL,
  order.by = c("FC_ExpMean", "FC_mean", "stat_statistic"),
  stat_fun = t.test
)
}
\arguments{
\item{data}{expression matrix}

\item{group.by}{vector}

\item{group.1}{character or vector}

\item{group.2}{character or vector}

\item{order.by}{order method}

\item{stat_fun}{stat function, used for 'stat_statistic'. Now only t.test is valid.}
}
\value{
geneList
}
\description{
Prepare genelist for GSEA
}
\examples{

}
