% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_regulon_specificity}
\alias{tl_regulon_specificity}
\title{calc Regulon Specificity Score (RSS) or named TFSS}
\usage{
tl_regulon_specificity(
  aucData,
  binData,
  group_by,
  groups = NULL,
  sig.auc.min = 0,
  sig.bin.min = 0.1,
  sig.rssz.min = 1,
  pseudocount = 0.01
)
}
\arguments{
\item{aucData}{auc scores matrix derived from SCENIC (e.g., 3.4_regulonAUC.Rds)}

\item{binData}{binary matrix}

\item{group_by}{a vector of clusters}

\item{groups}{clusters to be used}

\item{sig.auc.min, sig.bin.min, sig.rssz.min}{Regulon with AUCell score > sig.auc.min, Binary Activaty > sig.bin.min and RSSZ > sig.rssz.min were considered to be significant.}

\item{pseudocount}{used as numerator and denominator constant for calculation of fold change}
}
\value{
list of 1) matrix of group aggregated AUC scores, 2) matrix of Z-score normalized RSS, 3) data.frame of significant regulon
}
\description{
calc Regulon Specificity Score (RSS) or named TFSS
}
\examples{

}
