% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_RunPCAScore}
\alias{tl_RunPCAScore}
\title{a robust PCA for Seurat}
\usage{
tl_RunPCAScore(
  object,
  slot = "data",
  assay = NULL,
  reduction = "pca",
  use_all_genes = T,
  topn = 30,
  data.scale.max = 1,
  do.score.scale = T,
  score.scale.max = 1,
  reduction.name = "rpca",
  loadings = NULL
)
}
\arguments{
\item{object}{Seurat3 object}

\item{slot}{use which data to calc rpca}

\item{assay}{use which assay. NULL to use DefaultAssay(object)}

\item{reduction}{use which reduction to obtain top features.}

\item{use_all_genes}{get top n genes from all genes loadings, not just typically highly variable genes. This will call ProjectDim if inexist.}

\item{topn}{robust top n genes of loadings}

\item{data.scale.max}{scale parameter used in calculation of RPCA score. Set 0.99 or 0.95 as you like to exclude outliers.}

\item{do.score.scale}{scale robust scores}

\item{score.scale.max}{scale parameter used in calculation of RPCA score. Set 0.99 or 0.95 as you like to exclude outliers.}

\item{reduction.name}{new reduction name}

\item{loadings}{if set, use this loadings to calculate RPCA score}
}
\value{
updated Seurat3 object with a new pca_score reduction added
}
\description{
Based on a pre-computed dimensional reduction (typically calculated on a subset of genes and projects this onto the entire dataset (all genes).
Then generate a new rPCA reduction.
}
\examples{

}
