% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_RunSlingshot}
\alias{tl_RunSlingshot}
\title{a wrapper of slingshot}
\usage{
tl_RunSlingshot(
  object,
  reduction = "umap",
  group_by,
  start.clus,
  shrink = 0.75,
  ...
)
}
\arguments{
\item{object}{Seurat3 object}

\item{reduction}{character, umap or others}
}
\value{
Seurat3 object updated with ss_lineages, ss_Pseudotime and reductions$ss
}
\description{
a wrapper of slingshot
}
\examples{

}
