% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_transDim}
\alias{tl_transDim}
\title{Rotate or Flip reduction}
\usage{
tl_transDim(object, reduction = "umap", flip_x = F, flip_y = F, rotation = 0)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{reduction. eg. "umap"}

\item{flip_x}{do x flipping}

\item{flip_y}{do y flipping}

\item{rotation}{angle in degree to rotate}
}
\value{
updated Seurat
}
\description{
Rotate or Flip reduction
}
\examples{

}
