% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toCapitablize}
\alias{toCapitablize}
\title{wrapper for quick transform from human gene to mouse gene}
\usage{
toCapitablize(x, tolower = T)
}
\arguments{
\item{x}{vector}

\item{tolower}{do \code{tolower} first, then \code{Hmisc::capitalize}}
}
\value{
vector
}
\description{
wrapper for quick transform from human gene to mouse gene
}
\examples{

}
