% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toCheckedNames}
\alias{toCheckedNames}
\title{Converted to R check names}
\usage{
toCheckedNames(x, pattern = " |-|/", ...)
}
\arguments{
\item{x}{charaters}

\item{pattern}{regular expression, used in `gsub`}

\item{...}{other parameters used in `gsub`}
}
\value{
checked characters
}
\description{
R will modify some character of names (such as `-|/ `) to '.',
So for consitency, this function will help to get the checked names
}
\examples{

}
