% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{umi2tpm}
\alias{umi2tpm}
\title{normalize UMI matrix to TPM}
\usage{
umi2tpm(umi, scale.factor = 1e+06, margin = 2)
}
\arguments{
\item{umi}{umi matrix}

\item{scale.factor}{scale factor. default 1e6. recommend the mean of colSums(umi)}

\item{margin}{Cells in which margin}
}
\value{
a matrix of tpm values
}
\description{
normalize UMI matrix to TPM
}
\examples{

}
