% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniq_len}
\alias{uniq_len}
\title{Length of Unique values}
\usage{
uniq_len(x, ...)
}
\arguments{
\item{x}{a vector or a data frame or an array or \code{NULL}.}

\item{...}{other arguments passed to unique}
}
\value{
numeric
}
\description{
Length of Unique values
}
\examples{

}
