% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlist_depth}
\alias{unlist_depth}
\title{unlist by n times}
\usage{
unlist_depth(x, depth = NA, use.names = T)
}
\arguments{
\item{x}{list}

\item{depth}{n depth}

\item{use.names}{ogical. Should names be preserved?}
}
\value{
NULL or an expression or a vector of an appropriate mode to hold the list components.
}
\description{
unlist by n times
}
\examples{

}
