% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vectorSmooth}
\alias{vectorSmooth}
\title{smooth a vector}
\usage{
vectorSmooth(x, n)
}
\arguments{
\item{x}{a vector}

\item{n}{smooth over \code{n} items}
}
\value{
vector
}
\description{
smooth a vector
}
\examples{

}
