% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write.gmt}
\alias{write.gmt}
\title{write gmt file}
\usage{
write.gmt(gmt.list = NULL, desc.list = NULL, gmtfile = NULL)
}
\arguments{
\item{gmt.list}{geneset list}

\item{desc.list}{description list}

\item{gmtfile}{file name}
}
\description{
write gmt file
}
\examples{

}
