% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_clip}
\alias{write_clip}
\title{Write clipboard}
\usage{
write_clip(content, quoted = T, ...)
}
\arguments{
\item{content}{An object to be written to the system clipboard.}

\item{quoted}{quote each character}

\item{...}{other params passed to clipr::write_clip}
}
\value{
a string
}
\description{
A wrapper of clipr::write_clip.
Write a character vector to the system clipboard
}
\examples{

}
