% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_dataToJavaGSEA}
\alias{write_dataToJavaGSEA}
\title{write GSEA files}
\usage{
write_dataToJavaGSEA(
  data,
  group_by,
  groups = NULL,
  file_prefix = "gsea",
  add.sys.time = T
)
}
\arguments{
\item{data}{feature by cell matrix}

\item{group_by}{cell annotations}

\item{groups}{subset groups}

\item{file_prefix}{gsea file prefix}

\item{add.sys.time}{add time tag to gsea file name}
}
\value{
produce two files of .txt and .cls
}
\description{
write GSEA files
}
\examples{

}
