% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_genesToMeatascape}
\alias{write_genesToMeatascape}
\title{a helper function to write files for metascape analysis.}
\usage{
write_genesToMeatascape(genes, group_by, txt_file, top_n = NULL)
}
\arguments{
\item{genes}{a vector of genes (e.g. Symbols)}

\item{group_by}{a vector of gene annotation (e.g. Cluster). should be equal lenght of genes.}

\item{txt_file}{text filenamme}

\item{top_n}{only write top_n genes for each group.}
}
\value{

}
\description{
a helper function to write files for metascape analysis.
}
