% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_seuratToScenic}
\alias{write_seuratToScenic}
\title{write SCENIC data for in-house scenic-pureR}
\usage{
write_seuratToScenic(
  object,
  group_by,
  colVars,
  file,
  slot = "data",
  assay = "RNA"
)
}
\arguments{
\item{object}{seurat3 object}

\item{group_by}{vector of colnames}

\item{colVars}{colors vector}

\item{file}{save data to this file}

\item{slot}{fetch data}

\item{assay}{fetch data}
}
\value{
a saved file
}
\description{
write SCENIC data for in-house scenic-pureR
}
\examples{

}
