% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeDataFrameToClipboard}
\alias{writeDataFrameToClipboard}
\title{write data.frame to clipboard}
\usage{
writeDataFrameToClipboard(x, sep = "\\t", row.names = F, col.names = F, ...)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame.
    If not, it is attempted to coerce \code{x} to a data frame.}

\item{sep}{separator}

\item{row.names}{either a logical value indicating whether the row
    names of \code{x} are to be written along with \code{x}, or a
    character vector of row names to be written.}

\item{col.names}{either a logical value indicating whether the column
    names of \code{x} are to be written along with \code{x}, or a
    character vector of column names to be written.  See the section on
    \sQuote{CSV files} for the meaning of \code{col.names = NA}.}

\item{...}{other params passed to \code{\link{utils::write.table}}}
}
\value{

}
\description{
write data.frame to clipboard
}
\examples{

}
