% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConvertHomoGene}
\alias{ConvertHomoGene}
\title{convert genes symbols between organisms}
\source{
# genes, homology, orthology
http://www.informatics.jax.org/downloads/reports/index.html
https://gist.github.com/FloWuenne/f8fc922477df04c1642e9d8945c48d47
https://github.com/saeyslab/nichenetr/blob/master/R/supporting_functions.R
https://github.com/saeyslab/nichenetr/blob/master/R-raw/developing_package_light.R
}
\usage{
ConvertHomoGene(genes, orgA = "human", orgB = "mouse", verbose = F)
}
\arguments{
\item{genes}{genes to be converted}

\item{orgA}{from which organism, now human, mouse, macaque are supported}

\item{orgB}{to which organism, now human, mouse, macaque are supported}

\item{verbose}{report information}
}
\value{
data.frame of genes' mapping
}
\description{
using getLDS function of biomaRt to convert human genes to mouse genes or vice versa.
now only symbol is supported, which is used more frequently.
}
\examples{

}
