% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_tableHeatmap}
\alias{pl_tableHeatmap}
\title{Heatmap of a 2D-table}
\usage{
pl_tableHeatmap(
  tab,
  palette = "YlGn",
  palette_n = 5,
  min = 0,
  zero_as_na = TRUE,
  color_normalized = c("none", "row", "column"),
  color_scale_fun = I,
  number_normalized = c("none", "row", "column"),
  number_format = list(digits = 1, nsmall = 1),
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{tab}{a table object}

\item{palette}{palette name in RColorBrewer or a vector of gradient colors}

\item{palette_n}{interpolate number}

\item{min}{less than min will be setted to NA}

\item{zero_as_na}{coerce zero in tab to na}

\item{color_normalized}{color by normalized value}

\item{color_scale_fun}{color by scaled value with this function. Use I() to be asis.}

\item{number_normalized}{show number of normalized value}

\item{number_format}{arguments list for `format`}

\item{title}{plot title}

\item{xlab}{plot x lab}

\item{ylab}{plot y lab}
}
\value{
a ggplot object
}
\description{
Heatmap of a 2D-table
}
\examples{

}
