% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_crossTableEnrichment}
\alias{tl_crossTableEnrichment}
\title{statistics of cross table}
\usage{
tl_crossTableEnrichment(tab, p.adjust = T, heatmap = T)
}
\arguments{
\item{tab}{a table object}

\item{p.adjust}{adjust pvalue}

\item{heatmap}{plot heatmap}
}
\value{
a matrix of pvalues
}
\description{
statistics of cross table
}
