% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_goCompareCluster}
\alias{tl_goCompareCluster}
\title{A wrapper function for clusterprofiler enrichment analysis (compareCluster)}
\usage{
tl_goCompareCluster(
  degs_table,
  topn = 200,
  enrichFun = "clusterProfiler::enrichGO",
  OrgDb = "org.Hs.eg.db"
)
}
\arguments{
\item{degs_table}{deg results from FindAllMarkers}

\item{enrichFun}{enrich function name in clusterprofiler}

\item{OrgDb}{"org.Hs.eg.db" or "org.Mm.eg.db according to your organism}
}
\value{
a list of go and simplified go results
}
\description{
A wrapper function for clusterprofiler enrichment analysis (compareCluster)
}
\examples{

}
