% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl_statGene}
\alias{tl_statGene}
\title{calculate DEGs for two groups}
\usage{
tl_statGene(
  data,
  group.by,
  group.1,
  group.2 = NULL,
  FC_fun = function(x, pseudocount) {     Seurat::ExpMean(x + psuedocount) *
    log2(exp(1)) },
  stat_fun = stats::wilcox.test,
  pseudocount = 0
)
}
\arguments{
\item{data}{expression matrix}

\item{group.by}{vector}

\item{group.1}{character or vector}

\item{group.2}{character or vector}

\item{FC_fun}{Seurat::ExpMean or base::mean. log2(e) is equal to 1/log(2)}

\item{stat_fun}{wilcox.test or t.test}

\item{pseudocount}{avoid NA in calculation of fold change}
}
\value{
data.frame
}
\description{
useful for volcano plot
}
\examples{

}
