% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vlookup}
\alias{vlookup}
\title{vlookup with multiple keys supported}
\usage{
vlookup(df, x, keys = 1, select = NULL, collapse = "_", return.key = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{x}{values to lookup; data.frame of values per row if multiple keys used}

\item{keys}{colnames of \code{df} as key to search}

\item{select}{colname(s) to return}

\item{return.key}{whether to return keys}
}
\value{
vector or data.frame
}
\description{
vlookup with multiple keys supported
}
\examples{

}
